feat(intranet): move intranet service to a new intranet module
