+++ src/app/routes/api/v3/intranet/index.ts
@@ -1,0 +1,1 @@
+export { IntranetRouter } from './intranet.routes'
